tests: move safekeeper deletion tests into their own file
